total expenses visible in each categories

+++ backend/src/main/java/com/example/backend/controller/CategoryController.java
@@ -20,21 +20,21 @@ public class CategoryController {
 
     // Get all categories for a user
     @GetMapping
-    public ResponseEntity<List<CategoryResponseDTO>> getAllCategories(@PathVariable Integer userId) {
+    public ResponseEntity<List<CategoryResponseDTO>> getAllCategories(@PathVariable Long userId) {
         List<CategoryResponseDTO> categories = categoryService.getCategoriesByUser(userId);
         return ResponseEntity.ok(categories);
     }
 
     // Get a category by ID for a user
     @GetMapping("/{id}")
-    public ResponseEntity<CategoryResponseDTO> getCategoryById(@PathVariable Integer userId, @PathVariable Integer id) {
+    public ResponseEntity<CategoryResponseDTO> getCategoryById(@PathVariable Long userId, @PathVariable Long id) {
         CategoryResponseDTO category = categoryService.getCategoryByIdForUser(userId, id);
         return ResponseEntity.ok(category);
     }
 
     // Create a new category for a user
     @PostMapping
-    public ResponseEntity<CategoryResponseDTO> createCategory(@PathVariable Integer userId, @RequestBody Category category) {
+    public ResponseEntity<CategoryResponseDTO> createCategory(@PathVariable Long userId, @RequestBody Category category) {
         CategoryResponseDTO createdCategory = categoryService.createCategoryForUser(userId, category);
         return ResponseEntity.ok(createdCategory);
     }
@@ -42,8 +42,8 @@ public class CategoryController {
     // Update a category by ID for a user
     @PutMapping("/{id}")
     public ResponseEntity<CategoryResponseDTO> updateCategory(
-            @PathVariable Integer userId,
-            @PathVariable Integer id,
+            @PathVariable Long userId,
+            @PathVariable Long id,
             @RequestBody Category categoryDetails
     ) {
         CategoryResponseDTO updatedCategory = categoryService.updateCategoryForUser(userId, id, categoryDetails);
@@ -52,7 +52,7 @@ public class CategoryController {
 
     // Delete a category by ID for a user
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCategory(@PathVariable Integer userId, @PathVariable Integer id) {
+    public ResponseEntity<Void> deleteCategory(@PathVariable Long userId, @PathVariable Long id) {
         categoryService.deleteCategoryForUser(userId, id);
         return ResponseEntity.noContent().build();
     }
