+++ backend/src/main/java/com/example/backend/service/UserService.java
@@ -21,7 +21,7 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public User getUserById(Integer id) {
+    public User getUserById(Long id) {
         return userRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("User not found with id " + id));
     }
@@ -30,7 +30,7 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public User updateUser(Integer id, User userDetails) {
+    public User updateUser(Long id, User userDetails) {
         User user = getUserById(id);
         user.setName(userDetails.getName());
         user.setEmail(userDetails.getEmail());
@@ -38,7 +38,7 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteUser(Integer id) {
+    public void deleteUser(Long id) {
         User user = getUserById(id);
         userRepository.delete(user);
     }
